Re-architecture Phase 3: create legend-sdlc-core (L3); move updaters in
New module legend-sdlc-core, rooted at org.finos.legend.sdlc.core with
concern subpackages (core.entity/core.project/core.dependency/core.comparison;
no classes at the bare root). ProjectStructureUpdater moves from
org.finos.legend.sdlc.server.project (no bridge: no release ever shipped it
there) and ProjectConfigurationUpdater from
org.finos.legend.sdlc.server.domain.api.project (deprecated bridge subclass
retained) into org.finos.legend.sdlc.core.project. Resolves the Phase 2
carry-in on ProjectConfigurationUpdater placement: it is a configuration-delta
value object over L0 types and belongs at L3, not L4. Seam R1 (single
write-side dispatch point) moved verbatim.
