+++ docs/project-structure-migration.md
@@ -14,7 +14,8 @@ directly.
 | `org.finos.legend.sdlc.server.project.maven.*` | `org.finos.legend.sdlc.project.structure.maven.*` |
 | `org.finos.legend.sdlc.server.project.extension.ProjectStructureExtension` / `…Provider` / `UpdateProjectStructureExtension` (interfaces) | `org.finos.legend.sdlc.project.structure.extension.*`; deprecated bridge interfaces remain at the old FQNs |
 | `EntitySourceDirectory` nested in `ProjectStructure` | top-level `org.finos.legend.sdlc.project.structure.EntitySourceDirectory` |
-| `ProjectStructure.newUpdateBuilder(…)` / `ProjectStructure.UpdateBuilder` | `ProjectStructureUpdater.newUpdateBuilder(…)` / `ProjectStructureUpdater.UpdateBuilder` (still `org.finos.legend.sdlc.server.project`, in `legend-sdlc-server`; moves to the core module in Phase 3) |
+| `ProjectStructure.newUpdateBuilder(…)` / `ProjectStructure.UpdateBuilder` | `ProjectStructureUpdater.newUpdateBuilder(…)` / `ProjectStructureUpdater.UpdateBuilder` in `org.finos.legend.sdlc.core.project`, in **`legend-sdlc-core`** (Phase 3; it sat briefly in `org.finos.legend.sdlc.server.project` between the Phase 2 and Phase 3 commits, but no release shipped it there — no bridge) |
+| `org.finos.legend.sdlc.server.domain.api.project.ProjectConfigurationUpdater` in `legend-sdlc-server` | `org.finos.legend.sdlc.core.project.ProjectConfigurationUpdater` in **`legend-sdlc-core`**; a deprecated bridge subclass remains at the old FQN (note: the fluent `with*` methods return the relocated type) |
 | Structure code throws `LegendSDLCServerException` (JAX-RS `Status`) | throws `org.finos.legend.sdlc.error.LegendSDLCException` (int status code, same values); `LegendSDLCServerException` is its deprecated subclass |
 | `org.finos.legend.sdlc.server.project.ProjectFileAccessProvider` (and `ProjectFileOperation`, `ProjectFiles`, `ProjectPaths`, `AbstractFileAccessContext`, `CachingFileAccessContext`, `EmptyFileAccessContext`) in `legend-sdlc-project-files` | `org.finos.legend.sdlc.project.files.*` (same module; **no deprecated bridges** — update imports and recompile). The `SourceSpecification`/`WorkspaceSpecification` taxonomy keeps its `server.domain.api.*` packages until the backend-SPI phase decides its final home. |
