+++ legend-sdlc-core/src/test/java/org/finos/legend/sdlc/core/project/TestProjectConfigurationUpdater.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.sdlc.server.domain.api.project;
+package org.finos.legend.sdlc.core.project;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.StreamReadFeature;
