+++ legend-sdlc-server/src/test/java/org/finos/legend/sdlc/server/gitlab/api/TestGitLabEntityApiCharacterization.java
@@ -24,7 +24,7 @@ import org.finos.legend.sdlc.domain.model.revision.Revision;
 import org.finos.legend.sdlc.server.domain.api.entity.EntityAccessContext;
 import org.finos.legend.sdlc.server.domain.api.entity.EntityApi;
 import org.finos.legend.sdlc.server.domain.api.entity.EntityModificationContext;
-import org.finos.legend.sdlc.server.domain.api.project.ProjectConfigurationUpdater;
+import org.finos.legend.sdlc.core.project.ProjectConfigurationUpdater;
 import org.finos.legend.sdlc.server.domain.api.project.source.SourceSpecification;
 import org.finos.legend.sdlc.server.domain.api.project.source.WorkspaceSourceSpecification;
 import org.finos.legend.sdlc.server.domain.api.workspace.WorkspaceSpecification;
@@ -34,7 +34,7 @@ import org.finos.legend.sdlc.project.files.ProjectFileOperation;
 import org.finos.legend.sdlc.project.structure.EntitySourceDirectory;
 import org.finos.legend.sdlc.project.structure.ProjectStructure;
 import org.finos.legend.sdlc.server.project.InMemoryProjectFileAccessProvider;
-import org.finos.legend.sdlc.server.project.ProjectStructureUpdater;
+import org.finos.legend.sdlc.core.project.ProjectStructureUpdater;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
