+++ pom.xml
@@ -40,6 +40,7 @@
         <module>legend-sdlc-server-shared</module>
         <module>legend-sdlc-project-files</module>
         <module>legend-sdlc-project-structure</module>
+        <module>legend-sdlc-core</module>
         <module>legend-sdlc-server</module>
         <module>legend-sdlc-server-fs</module>
         <module>legend-sdlc-generation-shared</module>
@@ -495,6 +496,11 @@
             <dependency>
                 <groupId>org.finos.legend.sdlc</groupId>
                 <artifactId>legend-sdlc-project-structure</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.finos.legend.sdlc</groupId>
+                <artifactId>legend-sdlc-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
